wooo more daily points wip

+++ BreganTwitchBot.Domain/DTOs/Twitch/Commands/Points/GetPointsResponse.cs
@@ -9,7 +9,7 @@ namespace BreganTwitchBot.Domain.DTOs.Twitch.Commands.Points
     public class GetPointsResponse
     {
         public required string TwitchUsername { get; set; }
-        public required int Points { get; set; }
+        public required long Points { get; set; }
         public required string Position { get; set; }
     }
 }
